factor out user actions and improve nickname UX
