+++ frontend/src/actions/index.js
@@ -1,3 +1,4 @@
 export * from "./grid";
 export * from "./game";
 export * from "./messages";
+export * from "./user";
