+++ frontend/src/reducers/user.js
@@ -1,7 +1,8 @@
-import { SET_NICKNAME, REQUEST_SET_NICKNAME } from "../actions";
+import { SET_NICKNAME, REQUEST_SET_NICKNAME, SET_USERID } from "../actions";
 const initialState = {
   nickname: null,
-  nicknameRequested: false
+  nicknameRequested: false,
+  userId: null
 };
 
 export default (state = initialState, action) => {
@@ -10,6 +11,8 @@ export default (state = initialState, action) => {
       return { ...state, nickname: action.nickname };
     case REQUEST_SET_NICKNAME:
       return { ...state, nicknameRequested: true };
+    case SET_USERID:
+      return { ...state, userId: action.id };
     default:
       return state;
   }
