+++ frontend/src/sagas/user.js
@@ -1,3 +1,4 @@
+import jwt_decode from "jwt-decode";
 import { eventChannel } from "redux-saga";
 import { call, put, all, take } from "redux-saga/effects";
 import { putFrom } from "./index";
@@ -5,7 +6,8 @@ import {
   SENT_AUTH,
   SET_NICKNAME,
   REQUEST_SET_NICKNAME,
-  SET_TOKEN
+  SET_TOKEN,
+  setUserId
 } from "../actions";
 
 function* userFlow(socket) {
@@ -18,6 +20,8 @@ function* auth(socket) {
   if (localStorage.authToken) {
     socket.emit("auth", localStorage.authToken);
     yield put({ type: SENT_AUTH, new: false });
+    const { userId } = jwt_decode(localStorage.authToken);
+    yield put(setUserId(userId));
   } else {
     socket.emit("new auth");
     yield put({ type: SENT_AUTH, new: true });
